requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
mysql-connector-repackaged==0.3.1
python-dotenv==1.0.0
Werkzeug==2.3.6
pandas
numpy
matplotlib
pymysql
seaborn
scikit-learn
xgboost
joblib